requirements.txt
click==8.1.7
Faker==25.2.0
joblib==1.4.2
mysql-connector-python==8.4.0
nltk==3.8.1
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.5.15
scikit-learn==1.5.0
scipy==1.13.0
six==1.16.0
threadpoolctl==3.5.0
tqdm==4.66.4
tzdata==2024.1
